docs: replace content-control examples with one canonical tagged-inline-text (IT-1046)
The first pass shipped two examples (smart-fields, reusable-section) that read as demo decompositions: same primitives as demos/contract-templates, ~200 lines each, with markdown seeds and panel chrome. They didn't earn their place against the demo.

Replaced with one tight example, modelled on examples/editor/custom-ui/selection-capture: smallest readable form of the canonical create-find-update loop, with a README that distinguishes setup mutations from the teaching surface (selectByTag + text.setValue). Sets the template that future content-control examples copy.

- examples/document-api/content-controls/tagged-inline-text/ is the new single example.
- Old smart-fields/, reusable-section/, and the parent README.md removed.
- features/content-controls.mdx Next steps cards updated to point at the new example.
- Manifest and README rows replaced.
