--- examples/document-api/content-controls/reusable-section/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@superdoc-examples/content-controls-reusable-section",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "predev": "pnpm --filter superdoc build",
-    "dev": "vite",
-    "build": "tsc --noEmit && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "superdoc": "workspace:*"
-  },
-  "devDependencies": {
-    "typescript": "catalog:",
-    "vite": "catalog:"
-  }
-}
