--- examples/document-api/content-controls/smart-fields/vite.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite';
-
-export default defineConfig({
-  server: {
-    host: '127.0.0.1',
-  },
-});
