+++ examples/document-api/content-controls/tagged-inline-text/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@superdoc-examples/content-controls-smart-fields",
+  "name": "@superdoc-examples/content-controls-tagged-inline-text",
   "private": true,
   "type": "module",
   "scripts": {
