+++ examples/document-api/content-controls/tagged-inline-text/src/style.css
@@ -6,7 +6,6 @@
   --text-muted: #71717a;
   --accent: #1355ff;
   --accent-hover: #0f44cc;
-  --accent-soft: #eef3ff;
 }
 
 * { box-sizing: border-box; }
@@ -53,27 +52,28 @@ button { cursor: pointer; }
   line-height: 1.5;
 }
 
-.meta {
-  display: flex;
-  justify-content: space-between;
-  align-items: baseline;
-  padding: 4px 0;
+label { display: block; margin-bottom: 8px; }
+label span {
+  display: block;
   font-size: 12px;
   color: var(--text-muted);
+  margin-bottom: 4px;
 }
-.meta strong { color: var(--text); font-weight: 600; }
 
-.banner {
-  margin-top: 10px;
-  padding: 10px 12px;
-  border: 1px solid var(--accent);
-  background: var(--accent-soft);
-  border-radius: 6px;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
+input {
+  width: 100%;
+  height: 30px;
+  padding: 0 10px;
+  border: 1px solid var(--border);
+  border-radius: 4px;
+  background: var(--bg);
+  color: var(--text);
 }
-.banner-text { font-size: 12px; color: var(--text); }
+input:focus {
+  outline: 1px solid var(--accent);
+  outline-offset: -1px;
+  border-color: var(--accent);
+}
 
 .btn {
   height: 30px;
@@ -83,6 +83,7 @@ button { cursor: pointer; }
   border-radius: 4px;
   color: var(--text);
   width: 100%;
+  margin-top: 4px;
 }
 .btn.primary { background: var(--accent); border-color: var(--accent); color: #fff; }
 .btn.primary:hover:not(:disabled) {
