+++ examples/manifest.json
@@ -130,22 +130,12 @@
     "ci": true
   },
   {
-    "id": "document-api-content-controls-smart-fields",
-    "title": "Content controls: smart fields",
+    "id": "document-api-content-controls-tagged-inline-text",
+    "title": "Content controls: tagged inline text",
     "category": "Document API",
     "surface": "Content controls",
     "sourceRepo": "superdoc-dev/superdoc",
-    "sourcePath": "examples/document-api/content-controls/smart-fields",
-    "docs": "https://docs.superdoc.dev/document-api/features/content-controls",
-    "ci": false
-  },
-  {
-    "id": "document-api-content-controls-reusable-section",
-    "title": "Content controls: reusable section",
-    "category": "Document API",
-    "surface": "Content controls",
-    "sourceRepo": "superdoc-dev/superdoc",
-    "sourcePath": "examples/document-api/content-controls/reusable-section",
+    "sourcePath": "examples/document-api/content-controls/tagged-inline-text",
     "docs": "https://docs.superdoc.dev/document-api/features/content-controls",
     "ci": false
   },
